Darwin/Mac OS X fixes: sendfile() implementation, qmake stuff to compile it, arg(QThread::currentThread()) evals to arg(void *) so cast it to something safe
git-svn-id: http://svn.mythtv.org/svn/trunk@10140 7dbf422c-18fa-0310-86e9-fd20926502f2

+++ mythtv/libs/libmythupnp/httpserver.cpp
@@ -193,7 +193,7 @@ void HttpWorkerThread::StartWork( int nSocket )
 void  HttpWorkerThread::ProcessWork()
 {
     VERBOSE( VB_UPNP, QString( "HttpWorkerThread::ProcessWork:Begin( %1 ) socket=%2" )
-                                    .arg( QThread::currentThread() )
+                                    .arg( (long)QThread::currentThread() )
                                     .arg( m_nSocket ));
 
     bool                    bTimeout   = false;
@@ -294,7 +294,7 @@ void  HttpWorkerThread::ProcessWork()
     m_nSocket = 0;
 
     VERBOSE( VB_UPNP, QString( "HttpWorkerThread::ProcessWork:End( %1 )")
-                                    .arg( QThread::currentThread() ));
+                                    .arg( (long)QThread::currentThread() ));
 }
